7/19/23 Added Font.h, Font.cpp, edited most everything else - Ken

+++ Source/Engine/Renderer/Renderer.h
@@ -4,7 +4,6 @@
 
 namespace umbra 
 {
-	//void CreateWindow(const std::string& title, int width, int height);
 
 	class Renderer
 	{
@@ -19,9 +18,10 @@ namespace umbra
 			void BeginFrame();
 			void EndFrame();
 
-			void SetColor(int r, int g, int b, int a); //gets colors using int values
+			void SetColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a); //gets colors using int values
 			void DrawLine(int x1, int y1, int x2, int y2); //its LINE time!!
 			void DrawLine(float x1, float y1, float x2, float y2); 
+
 			void DrawPoint(int x, int y); //math forever haunts me
 			void DrawPoint(float x, float y);
